+++ src/test/java/de/is24/deadcode4j/analyzer/hibernateannotations/ClassWithDuplicatedTypeDef.java
@@ -1,0 +1,8 @@
+package de.is24.deadcode4j.analyzer.hibernateannotations;
+
+import org.hibernate.annotations.TypeDef;
+
+@SuppressWarnings("unused")
+@TypeDef(name = "aRandomType", typeClass = String.class)
+public class ClassWithDuplicatedTypeDef {
+}